Isolating contract-related deployment methods
