+++ packages/fuels/src/cli/commands/deploy/deployContracts.ts
@@ -4,11 +4,27 @@ import type { DeployContractOptions } from '@fuel-ts/contract';
 import { Contract } from '@fuel-ts/program';
 import { existsSync, readFileSync } from 'fs';
 
-import { setForcTomlProxyAddress, type ForcToml } from '../../config/forcUtils';
-import { debug } from '../../utils/logger';
+import {
+  getABIPath,
+  getBinaryPath,
+  getClosestForcTomlDir,
+  getContractCamelCase,
+  getContractName,
+  getStorageSlotsPath,
+  readForcToml,
+  setForcTomlProxyAddress,
+  type ForcToml,
+} from '../../config/forcUtils';
+import type { FuelsConfig, DeployedContract } from '../../types';
+import { debug, log } from '../../utils/logger';
 
+import { createWallet } from './createWallet';
+import { getDeployConfig } from './getDeployConfig';
 import { Src14OwnedProxy, Src14OwnedProxyFactory } from './proxy/types';
 
+/**
+ * Deploys one contract.
+ */
 export async function deployContract(
   wallet: WalletUnlocked,
   binaryPath: string,
@@ -112,4 +128,52 @@ export async function deployContract(
   setForcTomlProxyAddress(contractPath, proxyContractId);
 
   return proxyContractId;
+}
+
+/**
+ * Deploys all contracts.
+ */
+export async function deployContracts(config: FuelsConfig) {
+  const contracts: DeployedContract[] = [];
+
+  const wallet = await createWallet(config.providerUrl, config.privateKey);
+
+  log(`Deploying contracts to: ${wallet.provider.url}`);
+
+  const contractsLen = config.contracts.length;
+
+  for (let i = 0; i < contractsLen; i++) {
+    const contractPath = config.contracts[i];
+    const forcTomlPath = getClosestForcTomlDir(contractPath);
+    const binaryPath = getBinaryPath(contractPath, config);
+    const abiPath = getABIPath(contractPath, config);
+    const storageSlotsPath = getStorageSlotsPath(contractPath, config);
+    const projectName = getContractName(contractPath);
+    const contractName = getContractCamelCase(contractPath);
+    const tomlContents = readForcToml(forcTomlPath);
+    const deployConfig = await getDeployConfig(config.deployConfig, {
+      contracts: Array.from(contracts),
+      contractName,
+      contractPath,
+    });
+
+    const contractId = await deployContract(
+      wallet,
+      binaryPath,
+      abiPath,
+      storageSlotsPath,
+      deployConfig,
+      contractPath,
+      tomlContents
+    );
+
+    debug(`Contract deployed: ${projectName} - ${contractId}`);
+
+    contracts.push({
+      name: contractName,
+      contractId,
+    });
+  }
+
+  return contracts;
 }
